requirements.txt
google-generativeai>=0.3.0
google-cloud-firestore>=2.0.0
